+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `ui/Scroller` and `ui/VirtualList` not to scroll diagonally by flick
+
 ## [3.4.9] - 2020-10-30
 
 No significant changes.